requirements.txt
open_clip_torch~=2.0
clip-anytorch~=2.5.0
torch==1.12.1
transformers==4.23.1